feat: feature gate zksync os
