+++ Cargo.toml
@@ -22,7 +22,7 @@ strip = "debuginfo" # Automatically strip symbols from the binary.
 lto = "thin"        # Enable link-time optimization.
 
 [workspace.package]
-version = "0.6.9" # x-release-please-version
+version = "0.6.9"                                          # x-release-please-version
 edition = "2024"
 authors = ["The Matter Labs Team <hello@matterlabs.dev>"]
 homepage = "https://zksync.io/"
@@ -127,7 +127,7 @@ backon = "1.3.0"
 anvil_zksync_api_decl = { path = "crates/api_decl" }
 anvil_zksync_api_server = { path = "crates/api_server" }
 anvil_zksync_config = { path = "crates/config" }
-anvil_zksync_core = { path = "crates/core" }
+anvil_zksync_core = { path = "crates/core", default-features = false }
 anvil_zksync_l1_sidecar = { path = "crates/l1_sidecar" }
 anvil_zksync_types = { path = "crates/types" }
 anvil_zksync_common = { path = "crates/common" }
