+++ crates/core/Cargo.toml
@@ -26,12 +26,12 @@ zksync_error.workspace = true
 zksync-error-description.workspace = true
 
 # ZKsync OS dependencies
-forward_system.workspace = true
-basic_system.workspace = true
-zk_ee.workspace = true
-system_hooks.workspace = true
-zksync_os_api.workspace = true
-crypto.workspace = true
+forward_system = { workspace = true, optional = true }
+basic_system = { workspace = true, optional = true }
+zk_ee = { workspace = true, optional = true }
+system_hooks = { workspace = true, optional = true }
+zksync_os_api = { workspace = true, optional = true }
+crypto = { workspace = true, optional = true }
 
 anyhow.workspace = true
 tokio.workspace = true
@@ -71,3 +71,14 @@ httptest.workspace = true
 tempfile.workspace = true
 test-case.workspace = true
 backon.workspace = true
+
+[features]
+default = ["zksync-os"]
+zksync-os = [
+    "dep:forward_system",
+    "dep:basic_system",
+    "dep:zk_ee",
+    "dep:system_hooks",
+    "dep:zksync_os_api",
+    "dep:crypto",
+]
