+++ crates/core/src/node/keys.rs
@@ -10,7 +10,9 @@ impl StorageKeyLayout {
     pub fn get_nonce_key(&self, account: &Address) -> StorageKey {
         match self {
             StorageKeyLayout::Era => zksync_types::get_nonce_key(account),
-            StorageKeyLayout::ZKsyncOs => crate::node::zksync_os::zksync_os_get_nonce_key(account),
+            StorageKeyLayout::ZKsyncOs => {
+                crate::node::zksync_os::ZkSyncOSHelpers::get_nonce_key(account)
+            }
         }
     }
 
@@ -18,7 +20,7 @@ impl StorageKeyLayout {
         match self {
             StorageKeyLayout::Era => zksync_types::utils::storage_key_for_eth_balance(address),
             StorageKeyLayout::ZKsyncOs => {
-                crate::node::zksync_os::zksync_os_storage_key_for_eth_balance(address)
+                crate::node::zksync_os::ZkSyncOSHelpers::storage_key_for_eth_balance(address)
             }
         }
     }
